Index only well-formed account registrations from non-coinbase txs
Extract shared isWellFormedAccountRegistration() validator into
TransactionExtra and reuse it in both Core::check_tx_semantic (mempool
rejection) and Blockchain::pushTransaction (indexing gate). Coinbase
transactions are now skipped for registration indexing.
